fix calloc and memmove

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: jtanaka <jtanaka@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/09 03:48:56 by jtanaka           #+#    #+#             */
-/*   Updated: 2020/10/10 02:48:38 by jtanaka          ###   ########.fr       */
+/*   Updated: 2020/10/10 19:19:42 by jtanaka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,13 @@
 void	*ft_calloc(size_t nmemb, size_t size)
 {
 	unsigned char	*allocation;
-	int				i;
+	unsigned int	i;
 
 	allocation = malloc(nmemb * size);
 	if (allocation == NULL)
 		return (NULL);
 	i = 0;
-	while (i < size)
+	while (i < (unsigned int)size)
 		allocation[i] = 0;
 	return (void *)allocation;
 }
